# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 2
     * @maxLength 140
     */
  subject: string;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  message: string;
  /**
     * Honeypot field; legitimate users leave this blank
     * @maxLength 200
     */
  website?: string;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GenerateIdeasInputCountry = typeof GenerateIdeasInputCountry[keyof typeof GenerateIdeasInputCountry];


export const GenerateIdeasInputCountry = {
  AU: 'AU',
  NZ: 'NZ',
  UK: 'UK',
  US: 'US',
  CA: 'CA',
} as const;

export interface GenerateIdeasInput {
  /**
     * @minItems 1
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 80
     */
  interests: string[];
  /**
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 80
     */
  skills: string[];
  /** @maxLength 400 */
  passionsDetail: string;
  /** @maxLength 400 */
  contributionDetail: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  time: string;
  /** @maxLength 100 */
  outcome: string;
  /** @maxLength 100 */
  mode: string;
  kidFriendly: boolean;
  country?: GenerateIdeasInputCountry;
}

export interface GeneratedIdea {
  id: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  tag: string;
  /**
     * @minLength 4
     * @maxLength 100
     */
  title: string;
  /**
     * @minLength 100
     * @maxLength 1200
     */
  description: string;
  /**
     * @minLength 2
     * @maxLength 80
     */
  time: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  first: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  why: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  impact: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  potential: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  expand: string;
}

export interface IdeasResult {
  generationId: string;
  /**
     * @minItems 6
     * @maxItems 6
     */
  ideas: GeneratedIdea[];
}