# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 2
     * @maxLength 140
     */
  subject: string;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  message: string;
  /**
     * Honeypot field; legitimate users leave this blank
     * @maxLength 200
     */
  website?: string;
}