# lib/api-zod/src/generated/types/generateIdeasInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateIdeasInputCountry } from './generateIdeasInputCountry';

export interface GenerateIdeasInput {
  /**
     * @minItems 1
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 80
     */
  interests: string[];
  /**
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 80
     */
  skills: string[];
  /** @maxLength 400 */
  passionsDetail: string;
  /** @maxLength 400 */
  contributionDetail: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  time: string;
  /** @maxLength 100 */
  outcome: string;
  /** @maxLength 100 */
  mode: string;
  kidFriendly: boolean;
  country?: GenerateIdeasInputCountry;
}