# lib/api-zod/src/generated/types/generatedIdea.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedIdea {
  id: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  tag: string;
  /**
     * @minLength 4
     * @maxLength 100
     */
  title: string;
  /**
     * @minLength 100
     * @maxLength 1200
     */
  description: string;
  /**
     * @minLength 2
     * @maxLength 80
     */
  time: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  first: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  why: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  impact: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  potential: string;
  /**
     * @minLength 60
     * @maxLength 900
     */
  expand: string;
}