# lib/api-zod/src/generated/types/ideasResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedIdea } from './generatedIdea';

export interface IdeasResult {
  generationId: string;
  /**
     * @minItems 6
     * @maxItems 6
     */
  ideas: GeneratedIdea[];
}